se empieza a implementar el modulo contable , se empiezan a cambiar los datos contables de double por decima

+++ Domain/Entities/Descuento.cs
@@ -4,7 +4,7 @@
     {
         public Guid IdDescuento { get; set; }
         public DateTime FechaInicioVigencia { get; set; }
-        public int Porcentaje { get; set; }
+        public double Porcentaje { get; set; }
 
         public ICollection<Recibo> Recibos { get; set; }
     }
